+++ results.go
@@ -10,6 +10,8 @@ import (
 	"fmt"
 	"goatcli/output"
 	"os"
+	"strings"
+	"time"
 
 	"github.com/robert-kisteleki/goatapi"
 	"github.com/robert-kisteleki/goatapi/result"
@@ -36,10 +38,15 @@ type resultFlags struct {
 	limit        uint
 }
 
-// Implementation of the "result" subcommand. Parses command line flags
-// and interacts with goatAPI to apply those filters+options to fetch results
+// Implementation stub of the "result" subcommand, starting from command line args
 func commandResult(args []string) {
 	flags := parseResultArgs(args)
+	commandResultFromFlags(flags)
+}
+
+// Actual implementation of the "result" subcommand. Based on flags it
+// interacts with goatAPI to apply those filters+options to fetch results
+func commandResultFromFlags(flags *resultFlags) {
 	filter, options := processResultFlags(flags)
 
 	formatter := options["output"].(string)
@@ -58,6 +65,17 @@ func commandResult(args []string) {
 		defer flags.saveFile.Close()
 	}
 
+	if flagVerbose && flags.stream {
+		limits := ""
+		if flags.limit != 0 {
+			limits = fmt.Sprintf("%d ", flags.limit)
+		}
+		fmt.Printf("# Listening on the stream for %sresults, starting at %v\n",
+			limits,
+			time.Now().UTC(),
+		)
+	}
+
 	// most of the work is done by goatAPI
 	// we receive results as they come in, via a channel
 	results := make(chan result.AsyncResult)
@@ -67,12 +85,20 @@ func commandResult(args []string) {
 	output.Start(formatter)
 	for result := range results {
 		if result.Error != nil {
-			fmt.Fprintf(os.Stderr, "ERROR: %s\n", result.Error)
+			if strings.Contains(result.Error.Error(), "EOF") {
+				fmt.Fprintf(os.Stderr, "EOF\n")
+			} else {
+				fmt.Fprintf(os.Stderr, "ERROR: %s\n", result.Error)
+			}
 		} else {
 			output.Process(formatter, result.Result)
 		}
 	}
 	output.Finish(formatter)
+
+	if flagVerbose && flags.stream {
+		fmt.Printf("# Done listening to the stream at %v\n", time.Now().UTC())
+	}
 }
 
 // Process flags (filters & options), pass most of them on to goatAPI